# enhancement: Rule table optimisations