--- a/internal/test/testdata/engine/case_25.yaml
+++ b/internal/test/testdata/engine/case_25.yaml
@@ -6,7 +6,7 @@ inputs:
     {
       "requestId": "test",
       "actions": [
-        "create", # matches `acme_creator` wildcard resource and parent scope rule with `admin` role
+        "create", # matches `acme_creator` wildcard resource and parent scope rule with `employee` role
         "delete", # not defined in leaf scope, is in parent scope for matching (narrowed) derived role
         "view:public", # matches `acme_jr_admin` action wildcard and `employee_that_owns_the_record` in parent scope
         "deny", # allowed in target scope, no matching actions in parent scopes