--- a/internal/test/testdata/server/playground/test/pgt_case_00.yaml
+++ b/internal/test/testdata/server/playground/test/pgt_case_00.yaml
@@ -114,6 +114,49 @@ playgroundTest:
                                                       "components":
                                                         [
                                                           {
+                                                            "kind": "KIND_CONDITION",
+                                                          },
+                                                        ],
+                                                      "event":
+                                                        {
+                                                          "status": "STATUS_ACTIVATED",
+                                                          "result": true,
+                                                        },
+                                                    },
+                                                    {
+                                                      "components":
+                                                        [
+                                                          {
+                                                            "kind": "KIND_POLICY",
+                                                            "policy": "cerbos.resource.album_object.vdefault",
+                                                          },
+                                                          {
+                                                            "kind": "KIND_ACTION",
+                                                            "action": "view",
+                                                          },
+                                                          {
+                                                            "kind": "KIND_ROLE",
+                                                            "role": "user",
+                                                          },
+                                                          {
+                                                            "kind": "KIND_SCOPE",
+                                                            "scope": "",
+                                                          },
+                                                          {
+                                                            "kind": "KIND_RULE",
+                                                            "rule": "rule-001",
+                                                          },
+                                                        ],
+                                                      "event":
+                                                        {
+                                                          "message": "Condition not satisfied",
+                                                          "status": "STATUS_SKIPPED",
+                                                        },
+                                                    },
+                                                    {
+                                                      "components":
+                                                        [
+                                                          {
                                                             "kind": "KIND_POLICY",
                                                             "policy": "cerbos.resource.album_object.vdefault",
                                                           },