An effort to improve the performance of the rule table implementation.

* evaluations now only query the table once, and the result set is reused repeatedly (rather than using sub-queries to retrieve subsets), thus reducing the number of allocations per evaluation.
* introduced an index for rule retrieval in place of the full scan, including glob-key support.
* separated `resource` and `derived role` conditions and parameters to allow more granular caching and reduced evaluations.
* other misc. caches and optimisations.

#### `existing rule table` -> `optimised rule table` geomean comparisons:

- `sec/op`: -15%
- `allocs/op`: -22%
- `b/ops`: -20%

#### `main branch` -> `optimised rule table` geomean comparisons:

Run on the relevant, comparable subset of tests. Still not as performant as the old engine implementation, but a step closer:

- `sec/op`: +19%
- `allocs/op`: +7%
- `b/ops`: +8%

### `benchstat` full `sec/op` outputs for reference:

**existing ruletable -> optimised ruletable**

| Benchmark   | before.txt (sec/op) | after.txt (sec/op) | Change   |
|-------------|---------------------|--------------------|----------|
| case_00-12  | 366.4µ ± 3%         | 266.3µ ± 1%        | -27.30%  |
| case_01-12  | 367.8µ ± 0%         | 265.6µ ± 1%        | -27.80%  |
| case_02-12  | 1036.2µ ± 1%        | 942.1µ ± 1%        | -9.08%   |
| case_03-12  | 892.0µ ± 1%         | 782.2µ ± 3%        | -12.31%  |
| case_04-12  | 670.7µ ± 1%         | 564.6µ ± 1%        | -15.82%  |
| case_05-12  | 104.2µ ± 0%         | 105.1µ ± 1%        | +0.92%   |
| case_06-12  | 5.782µ ± 2%         | 5.576µ ± 1%        | -3.55%   |
| case_07-12  | 1.161m ± 1%         | 1.149m ± 2%        | -1.01%   |
| case_08-12  | 20.83µ ± 2%         | 12.10µ ± 1%        | -41.88%  |
| case_09-12  | 1.006m ± 2%         | 1.023m ± 0%        | +1.65%   |
| case_10-12  | 577.8µ ± 4%         | 594.9µ ± 0%        | +2.96%   |
| case_11-12  | 359.0µ ± 1%         | 363.7µ ± 1%        | +1.29%   |
| case_12-12  | 3.997µ ± 1%         | 4.204µ ± 0%        | +5.19%   |
| case_13-12  | 316.3µ ± 1%         | 341.9µ ± 1%        | +8.12%   |
| case_14-12  | 424.9µ ± 1%         | 344.3µ ± 2%        | -18.97%  |
| case_15-12  | 424.4µ ± 1%         | 343.2µ ± 1%        | -19.14%  |
| case_16-12  | 31.93µ ± 3%         | 24.16µ ± 1%        | -24.35%  |
| case_17-12  | 19.059µ ± 1%        | 8.627µ ± 1%        | -54.73%  |
| case_18-12  | 71.26µ ± 2%         | 65.86µ ± 1%        | -7.59%   |
| case_19-12  | 1.314m ± 3%         | 1.360m ± 0%        | +3.51%   |
| case_20-12  | 180.6µ ± 2%         | 179.3µ ± 0%        | ~        |
| case_21-12  | 235.4µ ± 3%         | 231.6µ ± 1%        | ~        |
| case_22-12  | 434.5µ ± 1%         | 213.5µ ± 1%        | -50.86%  |
| case_23-12  | 304.8µ ± 1%         | 288.5µ ± 1%        | -5.33%   |
| case_24-12  | 535.2µ ± 1%         | 394.6µ ± 1%        | -26.27%  |
| case_25-12  | 1111.4µ ± 1%        | 666.7µ ± 0%        | -40.02%  |
| case_26-12  | 84.32µ ± 0%         | 72.31µ ± 1%        | -14.24%  |
| case_27-12  | 518.5µ ± 0%         | 487.6µ ± 0%        | -5.97%   |
| case_28-12  | 337.8µ ± 0%         | 313.1µ ± 1%        | -7.31%   |
| Geomean     | 225.0µ              | 190.2µ             | -15.46%  |

**`main` branch -> optimised ruletable:**

| Benchmark  | main.txt (sec/op) | after-partial.txt (sec/op) | Change   |
|------------|-------------------|----------------------------|----------|
| case_00-12 | 204.0µ ± 0%       | 266.3µ ± 1%                | +30.58%  |
| case_01-12 | 203.2µ ± 1%       | 265.6µ ± 1%                | +30.70%  |
| case_02-12 | 809.1µ ± 3%       | 942.1µ ± 1%                | +16.44%  |
| case_03-12 | 601.6µ ± 0%       | 782.2µ ± 3%                | +30.02%  |
| case_04-12 | 492.7µ ± 0%       | 564.6µ ± 1%                | +14.58%  |
| case_05-12 | 103.8µ ± 0%       | 105.1µ ± 1%                | +1.29%   |
| case_06-12 | 6.011µ ± 1%       | 5.576µ ± 1%                | -7.23%   |
| case_07-12 | 1.170m ± 10%      | 1.149m ± 2%                | -1.77%   |
| case_08-12 | 9.581µ ± 0%       | 12.104µ ± 1%               | +26.33%  |
| case_09-12 | 462.4µ ± 1%       | 1022.5µ ± 0%               | +121.15% |
| case_10-12 | 368.0µ ± 0%       | 594.9µ ± 0%                | +61.67%  |
| case_11-12 | 214.5µ ± 3%       | 363.7µ ± 1%                | +69.57%  |
| case_12-12 | 6.459µ ± 0%       | 4.204µ ± 0%                | -34.91%  |
| case_13-12 | 312.0µ ± 2%       | 341.9µ ± 1%                | +9.60%   |
| case_14-12 | 257.8µ ± 0%       | 344.3µ ± 2%                | +33.56%  |
| case_15-12 | 256.0µ ± 3%       | 343.2µ ± 1%                | +34.06%  |
| case_16-12 | 10.26µ ± 1%       | 24.16µ ± 1%                | +135.36% |
| case_17-12 | 8.904µ ± 1%       | 8.627µ ± 1%                | -3.11%   |
| case_18-12 | 57.64µ ± 2%       | 65.86µ ± 1%                | +14.25%  |
| case_19-12 | 1.236m ± 0%       | 1.360m ± 0%                | +10.07%  |
| case_20-12 | 158.0µ ± 1%       | 179.3µ ± 0%                | +13.44%  |
| case_21-12 | 203.0µ ± 3%       | 231.6µ ± 1%                | +14.09%  |
| case_22-12 | 321.4µ ± 0%       | 213.5µ ± 1%                | -33.58%  |
| case_23-12 | 277.6µ ± 2%       | 288.5µ ± 1%                | +3.93%   |
| Geomean    | 144.0µ            | 171.7µ                     | +19.26%  |